Split combat module into a directory: types, algorithm, systems
The single combat.rs grew to ~1100 lines (mostly tests). Break it up
along its natural seams:

- combat/types.rs (36 lines): UnitSnapshot, ResolveAction, CombatDeltas.
- combat/algorithm.rs (481 lines): resolve_movement_pure, the three
  per-phase helpers, and the eight pure-function tests.
- combat/systems.rs (569 lines): the three ECS wrapper systems and
  their six integration tests.
- combat.rs (13 lines): module declarations and a top-level pub use of
  the three systems main.rs depends on.

No behavior change. Public API (combat::{cleanup_dead_units,
resolve_movement, resolve_ranged_attacks}) is preserved via re-export.
Tests stay co-located with the code they exercise.
